Refactor: move journal and tasks pages
