+++ lib/pages/journal/journal_page.dart
