+++ lib/routes/task_routes.dart
@@ -1,7 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:lotti/pages/create/create_task_page.dart';
 import 'package:lotti/pages/journal/entry_details_page.dart';
-import 'package:lotti/pages/tasks_page.dart';
+import 'package:lotti/pages/tasks/tasks_page.dart';
 
 const AutoRoute taskRoutes = AutoRoute(
   path: 'tasks',
